+++ config/settings/local.py
@@ -87,6 +87,7 @@ CELERY_TASK_EAGER_PROPAGATES = True
 # Local App Settings
 # ------------------------------------------------------------------------------
 
+PROJECTROLES_ALLOW_LOCAL_USERS = False
 
 # Plugin settings
 ENABLED_BACKEND_PLUGINS = [