+++ config/settings/local_target.py
@@ -32,3 +32,6 @@ PROJECTROLES_SITE_MODE = 'TARGET'
 
 # Username of default admin for when regular users cannot be assigned to a task
 PROJECTROLES_DEFAULT_ADMIN = 'admin_target'
+
+# Allowing local users by default when developing target site locally
+PROJECTROLES_ALLOW_LOCAL_USERS = True